check convergence criterion depending on the expected solver type
